CoWinner:Make NodeJS ready: Export, UserAgent
To allow use of the module from nodejs, export the entry function
wrt the core logic of this module.

It was noticed that the CoWin API server is blocking fetch requests,
unless a user agent is specified. (Not sure why the default UserAgent
supposed to be added by node-fetch, is not working; need to check
later). So added a unique one wrt this module.