NodeJS: Initial skeleton to call into cowinner
As fetch is not defined in nodejs by default so node-fetch module
is used to provide support for same.

Inturn when using require to bring in the same, dont use const or
var, so that its accessible across modules, else it will be accessible
only in the module where its brought in through require.

This also avoids the need to conditionally bring in fetch in cowinner
module by checking for typeof(fetch) === 'undefined'.