Helper: Deploy to GitHub Pages - Project site
This allows one to deploy a custom subfolder of the project as the
project's online site on github pages hosting service.

NOTE: In this case that subfolder is hardcoded as browser (as that
is where the browser specific web app code is).

Remember to select the branch created here gh-pages in the project's
setting->pages page on github.