CoWinner: Multiple instances of a Vaccine Center
As the same vaccine center may provide vaccination for different
vaccines to different age groups, so there could be multiple
instances of the same vaccine center in the database, this is
accounted for now, by making the object storing data about any
given vaccine center into a array.

NodeJSVersion of the vaccheck program updated to handle the same
now.