MBrowserVer: Multiple VC instances, dblookup_vcs
The modular browser based logic has been updated to work with the
new db, which allows multiple instances of the same vaccine center
to exist in it.

Also it uses the dblookup_vaccenters logic, so that vc filtering
is consolidated at a single place in it. So both Browser based
and NodeJS based versions share most of the logics.