MBrowserVer: Cleanup view to some extent
Added a viewport meta to ensure that the table and input elements
dont look too small and out of sync with other elements on mobile.

Moved some of the web elements and text messages.

Added the search parameter related elements into a  table.

Realised that even thou I was passing a element to tbl helpers,
I wasnt limitting my search for table within this element. So
for now have added a index to tbl helpers, so that it works on
ith table in the page.