CoWinner: DBLookUp filter on specified state(s)
As db has been made persistant within a run session, so it can
accumulate data across multiple states, so one needs to filter for
the specific state the user is currently interested in as part of
the dblookup logic, the same is handled now.

In order to achieve the same, now the state(s) being searched for
is also stored as part of the db.

MBrowserVer has been updated to pass the state(s) array using db,
rather than passing it directly to dbget_states. NOTE: Even if
it was passed as a argument to dbget_states, it inturn would update
the db wrt the list of states the user is interested in, so that
later when dblookup is called, the same can be used for filtering
other states' that may be in the db.