CoWinner: Allow caching of vaccenters across multiple dates
As users may want to check for slot availability in the same state
across different dates, the prev caching logic wouldnt have worked
with reducing the server load beyond a limit, because it would have
allowed vac centers data to be saved for only one of the dates at
any time.

So this new logic allows VCs data to be saved wrt multiple dates
in the db. Thus should lead to better experience for user (less
waiting time) as well as the server (a more responsible client :)

This is the initial skeleton wrt achieving the same.