CoWinner: Update ver in UserAgent, Fix time display
As logic was updated in prev commit to enable caching for multiple
days wrt states, so the version embedded in the UserAgent sent
as part of the request to api server, as been updated.

During the change to support multiple days wrt any given state,
the way fetch time was stored, had changed. I had forgotten to
update the logic to account for same, wrt the time shown to user.
THis is fixed now.