MBrowserVer: Add Auto Search logic
Allow user to start or stop a auto search, which if enabled, will
keep triggering the search automatically at periodic intervals,
currently set to once every 10 minutes.

What ever is the currently selected State,Date and Vaccine will be
used as the parameters for the auto triggered search.

NOTE: User requires to trigger a manual search for the search
parameters to be picked up, after they have been set/changed.
Will be changing this to be automatic later.