MBrowserVer: Pickup latest search params always
Now irrespective of whether search btn is clicked or Auto btn is
clicked, the search params as set by user is picked before the
search is triggered.

NOTE: However if user changes the search parameters after auto
search is triggered, it wont be picked up, One will have to stop
and start the auto search.