MBrowserVer: bNotifyMode and bNotifyMe
Based on whether search is trigger by user clicking search button
or based on auto repeating search based triggering of search logic
bNotifyMode is setup, so that notify_user logic will be only triggered
for auto event (as user may not be actively looking at the program)
and not search btn click event (bcas user will be already looking at
the program, so no need to explicitly notify user).

Based on clicking of the Notify button, bNotifyMe is toggled as
required. Also the text on Notify button is changed appropriately.

ONly if both of these boolean flags are true, the notification
logic is triggered.