MBrowserVer: More clearer Notification
Previously as part of initial skeleton the full data that needs to
be shown to the user was put into the title, which makes it difficult
to see easily in some GUI systems (ubuntu one couldnt see towards
end of long titles in notification panel).

So now part of the data is put into title and the other part into
body of the notification, so that full data can be seen. Also body
will use a smaller font, so even more data can be fitted if reqd.