CoWinner: strlist_findindex, vcs_forweek, dbget_vcs
Added a generic strlist_findindex to find a specified string in
a given list/array of strings. The same is inturn used to filter
for state to work with wrt fetch and lookup logics.

Renamed dbget_vaccenters_fordistrict to dbget_vaccenters_forweek,
because it gets all vaccenters in a given district for upto a
week, provided the data is available on cowin system at the time
of query.

Renamed dbget_state_vcs to dbget_vcs, because it is being
made generic to fetch either all VCs with slots for a given date
with in a state across all its districts, OR ELSE to get all VCs
with slots for a given district for upto a full week (provided
the same data is available on cowin system at the time of query).

	The logic is not yet implemented.