CoWinner: STATE_1DAY default, CacheCheck handle s_type
Now if the db passed to dbget_vcs doesnt contain the s_type specified,
then it will be defaulted to STATE_1DAY.

The function that checks for cache or fresh, now handles the s_type
wrt caching or not, instead of dbget_vcs. This ensures that there
is a valid timestamp for Non Cached accesses of DISTRICT_1WEEK,...

These two changes also ensure that we dont end up in statedone
callback, without time stamp info in db.