CoWinner:MultiLvlCache: Generic CacheNotFresh, GetStates
Changed CacheNotFresh to be a simple and generic logic, so that it
can be used to check cache time wrt data at different levels in the
db or ...

GetStates updated to use caching to avoid querying server to fetch
list of states. Also if there is a exception when fetching then
cache time is reset to undefined, so that the next call to GetStates
will query the server.