CoWinner:MultiLvlCaching: Add caching to GetVCs4Week
GetVCs4Week maintains cache time at TheDistrict.time, while
GetVCs4Date maintains cache time at TheDistrict.TheDate.time.

So also the timeKey argument support added to CacheNotFresh is not
used for now, as for both GetVCs independent cache times are used.

Technically/Logically CoWin server should be providing the same
data for dates which overlap between fetching VCs data for a specific
date and fetching for a specific week, but for now this is not
optimised for as the fetch is triggered using different apis and
as there is some changes in data structure of the returned data.

	Its a different thing, that I am already adjusting the
	returned data such that in both cases, the VCs related
	data follows a similar structure within db.

	But for now caching is not considered/shared across fetch
	types, because we cant selectively fetch data for specifc
	days within the week (next 7 days) api. And there is a
	limit on number of api calls to the server in a given
	short time, so simulating week api call with multiple
	day api calls cant be done.