BrowserV: select helper, state->district
Added helpers to clear a select of its options, as well as to
populate a select with options picked from a passed object.
The helper for populating, also adds a predefined ANY option
to the passed select.

Added initial skeleton to trap a state change and inturn try
populate the district select with districts in that state.

TODO: For now its hard coded to state with id 17. Also the logic
doesnt try to populate the db, before using the same wrt states
and districts in the selected state.