CoWinner: Shared VC data across queries, so avoid duplicates
As the db caches and shares the VC data across State1Day and
District1Week queries, so one needs to ensure that duplicate
entries are not added into db.

Also one needs to ensure that the data in the db, is updated with
new values from newer queries.

This is handled now.

NOTE: This also assumes that server will not drop a VC after it has
been mentioned for a given date, even when the availability at that
VC goes down to 0. Currently CoWIn server is showing this behaviour
so this is ok.

However if in future the server, drops sending data for VCs which no
longer have slots available, then one will have to either update
the logic and or rather avoid sharing the data between different
search types.