CoWinner:DB: Shared but overwritten with latest
Even thou the DB shares the VC data across State1Day and District1Week
queries, instead of trying to check for duplicates when new search
results wrt same state-district-date are fetched, Now it replaces the
old dataset with a new dataset.

This should avoid the duplicate issue. So also the log in duplicate
check is enabled now, so we can see if it triggers (which shouldnt
occur, unless the server data itself has duplicate entries).