CoWinner: CacheTime arg for CacheNotFresh, Regions longer caching
CacheNotFresh now supports a cacheTime argument, which is used to
decide how long the corresponding data should be cached.

Inturn the same is used to cache regions (i.e states list and dists
list) for a longer time of around 1 hour. While the VCs data is
cached for upto 5 minutes.